Several changes to get up and running.
This is a stable configuration for GCN controllers using the 3bb24e218b61eadfb60ef6813b200da4f8299abd
commit from NintendoSpy (an arbitrary commit on master) on an Arduino Uno.

The changes include:

 * Some package updates. I don't remember why these were necessary, and for this I am sorry.
 * win.webContents.openDevTools() added to main/index.js, so a giant debug window pops up when starting. This can be removed.
 * stick.range was giving me trouble, so I removed it and added the range explicitly to stick objects in display_interface. IIRC this was the only actual bugfix, as range isn't correctly set in el.dataset (I think it is in custom skins?) so the sticks werent working despite everything else being alright.
 * some minor modifications to static/skins/gcn/skin.json because I needed to adjust everything to look right.

Overall this is a debug branch, but it has worked reliably for me for probably thirty hours of playing smash!